private/apigen/example: Fix client TS generated file
We introduced some changes in 882c9d64e4f1bbdaac62c90f3cbe04bb6cad00ee
and one reviewer spot an issue which was fixed, but I forgot to
regenerate the files and the generated Typescript client got stale.

This commit regenerate the example code, so the generated code is align
with the current generator version.

Change-Id: Idc023adc9c6e230578f9e77e492126355b006203