cmd/tools: add is-valid-sj1-blob utility
Takes a list of filenames, and evaluates each to see if it appears to be
a valid sj1 blob. (Any arbitrary blob is a valid FormatV0 blob, and we
can't tell anything helpful about the blob in that case, so we don't
bother checking for that.)

If the blob does not appear to be valid, output diagnostic information.
If it does appear to be valid, show the blob's correct filename inside a
storage node blobs directory.

Change-Id: I2c1cdde16c3b8c1a9953f4a0a2df816869f27409