satellite/admin/back-office: Don't send project ID
The project ID must never be sent out from the server because old
projects used that ID as salt for creating access grants.

We must always expose the public ID and the server must retrieve the ID
when the public ID is sent in to the server and needs it for performing
some DB queries.

Change-Id: I4a9d94049ef84f578b62907827d6c09dfd1db4db