cmd/storagenode: development Docker images for testing
This patch restores the Storagenode Docker image for development purposes.

This is not replacement of the Storagenode image of the end users (https://github.com/storj/storagenode-docker).

To avoid confusions, it's NOT pushed to Docker Registry, only to the development registry (img.dev.storj.io).

Image supposed to be compatible with all environments (storj-up, Kubernetes environments, ...),
but not public network (helper environment variables are intentionally missing).

Change-Id: I15d23d61c74b23bc6589351996c244d36408a699