satellite/repair: assign lower health to segments with pieces out of placement
Segments with pieces out of placement (which we'll call "POPs" here for
brevity) need to be repaired as soon as possible, even if they are
otherwise pretty healthy.

We don't want to say that _all_ segments with POPs are higher priority
than _all_ segments without POPs, but it should take a pretty severe
danger to make a segment with fewer POPs be prioritized before a segment
with more.

This change accomplishes this requirement along with a wholesale
refitting of the segment decay model, in hopes of addressing concerns
that the calculation and its output are too opaque and too hard to
understand. The new model, its appropriateness, and its application are
explained in much more detail than they were in the old code. As a nice
bonus, this model gives more reasonable health values when minPieces is
very low.

See
https://github.com/storj/datascience/blob/6aa8d85/repair_and_durability/repairPriority/hypergeo.ipynb
for more discussion of the mathematics of the model.

Change-Id: I3fd541834a37a253a08ded60ef2d475244653dd5