satellite/admin: Return project public ID in GET user
Return if the user is in the paid tier and the public ID of the projects
when getting the user information.

Showing the paid tier we can know in which tier the user is when we have
support tickets that require to change their tier.

Some times the solution engineers open requests to change project limits
and they inform the public project ID. Without this we must search
manually in the DB which project is when the user's account has more
than one.

Change-Id: I47813d443b3ca43179e906e30a79a5f511367d6a