satellite/admin: Fix user account delete
Delete a user's account that was member of foreign projects failed
because the endpoint was getting the list of projects that the account
is member which implies owned project, but also the foreign projects
that's member of.

This commit fix the issue calling the right method to only get the list
of owned projects to check that they are zero before proceeding and add
a test to verify it.

Change-Id: I0071394b1b26366bad8dfff3be1fc6edd418ac85