satellite/nodeselection: extend Placement with Selector
This extension of the placement makes it possible to define the selection types for uploads (per placement).

Each placement can use different selection mechanism.

The available implementations are the same as before:

  1. select a random node
  2. select a random subnet, and a random node from the subnet

Patch looks big, but the key part is the following two files:
 * satellite/nodeselection/selector.go
 * satellite/nodeselection/state.go

Which also requires some small changes in placement.go.

And the long tail is just unit test updates....

Change-Id: I0a814299247c21d06ef0dc73b5bbb40a9d758648