private/server: reject invalid connections
drpcmigrate.ListenMux is a Mux for raw data: it can route new incomming connections to the specified net.Listener instance (based on prefix bytes).

It also has a default net.Listener which is used if no other prefixed listener matched the connection.

But we don't handle the connections from the default listeners, therefore they are open forever.

This patch starts goroutines to close these connections.

Change-Id: Ic4d7d3d2f88a5487156da74ca28679a432b0ce1f