cmd/storagenode: revamp forget-satellite subcommand
Previously, the forget-satellite command had no direct connection
with the main storagenode process (i.e. the node is unaware that a
satellite is being forgotten until it is restarted).

The new forget-satellite sends a drpc request to the storagenode
to initiate the satellite cleanup.

Updates https://github.com/storj/storj/issues/6541

Change-Id: Ia207722f5e11c1a56f0a929afa21ac178dd713aa