satellite/satellitedb: handle simultaneous inserts to reputations table
It can happen that multiple processes want to suspend, unsuspend, or
disqualify a node at the same time. (Mostly this only happens in test
environments, where Postgres and Cockroach tests run concurrently, but
we should still handle the case in prod.)

When this happens, before this change, it was possible for both
processes to get an ErrNoRows error, indicating the node in question had
no row in the reputations table yet. Then both processes would try to
insert the same node into the table simultaneously, leading to one of
them getting a constraint violation error.

When the action we are taking is only inserting an empty reputations row
with an empty audit history, resolving the conflict is straightforward:
leave the existing entry there and do not insert any new row. This is
accomplished with the ON CONFLICT DO NOTHING clause being added here to
the INSERT statements.

This should resolve potential flakiness in
TestDBDisqualificationSuspension, TestNodeSuspended,
TestAuditSuspendBasic, TestSuspendedNodesFailGracefulExit,
TestAuditSuspendDQDisabled, TestAuditSuspendExceedGracePeriod,
TestDisqualifiedNodeRemainsDisqualified,
TestDisqualifiedNodesGetNoDownload, and
TestDisqualifiedNodesGetNoUpload.

Refs: https://github.com/storj/storj/issues/5496
Change-Id: I7bf722cbc89d3e18d300d5c4ec945c9ec3bfc9cf