satellite/console: feature flag to toggle v2 as default UI
This change adds a flag to toggle between serving the v1 or the v2 app
on the root path, especially making it the default satellite UI. It is
false by default, maintaining the current behavior, i.e. v1 app on root
path and v2 app on root path/v2.

Issue: #6681

Change-Id: Ifd48175e3c9f00d6187187f572f0c16fd30e09b6