satellite/admin/back-office/ui: Hook feature flags
Hook all the UI to the feature flags returned by the API /settings
endpoint.

The UI shows and hide access to views and dialogs considering if the
corresponding feature is enabled or disabled.

This commit also update the setting endpoint to enable the feature that
are now available because they were merged and add a new feature to
distinguish between list users' accounts and search for a user account,
restoring back the accounts views to what it was before introducing the
find account by e-mail and moving that one to new view.

Change-Id: Ief764dfec24ccc8c8e144dbcb2a61cbaa914118c