satellite/nodeselection: YAML based placement configuration
Existing placement configuration proven to be hard to read.

And we also added more functionalities to Placement (like selectors or invariants).

It's time to use external file, which is easier to read (and easier to maintain the same values for multiple microservice instances).

Legacy configuration loading will be removed after 1-2 releases.

Change-Id: I8ef4c001b60fb96f04a34eb3acc370349f620dd4