storagenode/payouts: show historical payouts data for old/untrusted satellites
Previously, payout history fetches data for only satellites
found in the trust cache. Which means that once a satellite
is not longer trusted, there will be no payout information
for the satellite and previous payout information were
not displayed.

With this change, we include payout history for untrusted
satellites. If there is no record of the satellite URL in
the satelliteDB (although we might fix that in this ticket
https://github.com/storj/storj/issues/6574\), we display
the satelliteID as the satelliteURL for untrusted satellites.

Fixes https://github.com/storj/storj/issues/6574

Change-Id: I4b20abe1abccb5d787fa0317f9d374cc563b7229