web/satellite/v2: fix unauthenticated navigation
This change fixes an issue where navigating directly to a page when
unauthenticated will show the page for a while before navigating to the
login page, by introducing a full screen loader as seen in the V1 app.
It also moves certain page setup calls (user, projects etc.) to App.vue
so as not to replicate the login-redirect logic in more places. This
also helps avoid calling for setup data that have already been fetched.
Finally, navigation directly to the authentication pages is prevented.

Issue: #6699

Change-Id: I3b5f8570c7d430f2afc5b0366f0e32852e4d1ae0