satellite/admin: fix set to defaults for project limtis
This commit fixes the following stuff related with the project limits
update.
- The client side allows the user setting back a field to blank after a
  value was inputted without sending any value to the server for that
  specific field. Previously, an invalid value was sent.
- The API server allows setting `rate`, `burst`, and `buckets` to `nil`,
  which means that the defaults will be applied. Previously, it wasn't
  possible.

The commit also stop exposing the update limit endpoint under the POST
method and only exposing it by PUT. There wasn't any benefit on exposing
with both methods and remove any possible confusion, which was evident
because the README documented one, while the client side implementation
used the other.

There several changes which are breaking long lines of code in several,
this cause noise, but I accepted to commit them because they were
automatically done by my editor and they look more readable to me. Sorry
for that!

Change-Id: I541e5613aee4dfec0bb19cd4204d07ede2bc720d