storagenode: trash with Bloom filter's timestamp
If a given bloom filter was received at a specific date, and a piece uploaded
before that date is not matching it, this means the satellite already expects
the piece to be trashed. Yet given that the garbage collection procedure may
take a lot of time, some pieces will effectively stay on a node for far longer
than the 7 days expected by the satellite. This patch corrects this behavior.

Change-Id: I7854c7a7e6e19e6a439ec2d58c5a486f421b27b0