satellite/audit: much stricter audit transfer speed reqs
With MinBytesPerSecond's old value of 128B, we were allowing multiple
hours for transfer of a maximum-size piece (~2MiB) during
reverifications. We do want to be generous until it has been proven that
something needs to be stricter, but we have hit this point. Hours is
much too long.

Instead, this new value of 150kB is still reasonably generous while
limiting transfer time for a full-size piece to around 15s.

In the case of downloading a single share during verification, the
MinDownloadTimeout will apply and allow for minor network hiccups. We
shorten that time here as well, from 5m to 15s. This should allow a much
higher rate of audit completion in the presence of very slow nodes.

Refs: https://github.com/storj/infra/issues/5390
Change-Id: I7402ebd91b2115b379e7671d6b7455d3b0eae1fb