satellite/metainfo: don't use metabase.ListObjects for unversioned buckets
This is workaround for metabase.ListObject performance issues.
Earlier IterateObjectsAllVersionsWithStatus was changed to return
results in descending order and regular listing was changed to use
ListObjects method. With this change IterateObjectsAllVersionsWithStatus is reverted under IterateObjectsAllVersionsWithStatusAscending name and
is used to do regular (only latest objects) listing for unversioned
buckets and for pending objects listing. Versioned buckets will be still
listed with metabase.ListObject.

Change-Id: I49b57edc1f7e352c035f602ac71048e8461e0632