satellite/nodeselection: file based node-id filter for selectors
We already use filter in the placement configuration, but that applies to all operations (upload/download/repair).

This patch introduce selector(=upload) only usage of the filters, which can adjust which nodes should be used for selection.

Also, a helper is added to load nodeId list from file.

This can be useful during maintencance of private clusters, as certain nodes can be included/excluded temporary with a list.

Change-Id: Iec83c0864a5fd5a580ecaa654e5e2601c3096dd1