storagenode: display actual disk space when allocated is higher than actual
For cases when the allocated disk space is higher than the actual
disk space, the dashboard will display the actual disk space.

Also displays the min(free, actual free space on disk).
This partially fixes the largely reported disk space discrepancies
where the values displayed on the dashboard does not match
the actual disk values.

This shouldn't be a problem for the satellite site since the
node already reports the min(free for storj, actual free space on disk)

Change-Id: I537cd1e653394aab843be27e3f4c143fed6f6823