web/satellite/v2: use onboarding step field
This change updates the account setup dialog to use the onboarding_step
field of user_settings to track setup progress. Enum values are also
added for the steps of the new onboarding stepper. Finally the call to
setup stripe has been removed from the setup dialog since it is already
called in the overall app scope.

Issue: #6721

Change-Id: I82cc5458cb699aa31ac0111fa0439b87e2bf4c87