web/satellite/v2: fix otp paste issue
This change fixes an issue where pasting an OTP code with extra text in
the OTP field will paste all that text instead of trimming it to the
required length. Non-number characters are now rejected.
It also fixes an issue where signup success redirect will fail because
of a malformed URL. Finally, more route cleanups are done and a minor
change has been made to the MFA loading UX.

Issue: #6751

Change-Id: I1bbc73e3eb476a2de022bd92ac66b3be4bd9d7fc