storagenode/{piece,blob}store: avoid extra trash checks
We check to see if missing pieces/blobs are in the trash so that we can
get an idea via metrics if things are going into the trash incorrectly.
However, we've been doing that twice for every time we get a not-found
error, which is unnecessary.

The first check happens in the blobstore. We currently use that to
increment a metric counter, then we throw the information away.
Afterward, at the piecestore endpoint layer, we check again, then try to
restore the blob out of the trash if possible.

Instead of doing that, we will remove the check from the blobstore, then
increment the same metric counter from the piecestore side if
appropriate.

Change-Id: I4513baf5a98c1df11d0d903ddf6658d024176303