storagenode/blobstore: correct naming piece->blob
The blobstore layer isn't supposed to know about "pieces". It deals only
in binary blobs, addressed by namespaces and keys. This change corrects
some comments and a few method names to reflect this. Maybe it doesn't
really matter in the long run, but I find it easier to understand this
code when the layers make sense and have a clear separation of duties.

Change-Id: I3c12e9a479c98fd691358c9313f760a62df749ae