storagenode/blobstore: store trash under per-day directories
This will allow the trash filewalker to traverse a (generally) much
smaller directory space when cleaning up expired trash.

The chief cost is that we need to look in more places (~7) when checking
whether a piece is in the trash. However, since we only need to pay that
price when the node is about to return a Not Found error, that seems
like a reasonable cost to pay.

Trash that exists from before will be migrated to this structure the
first time a node runs with this code. To keep it fast, we put all
existing trash in the current-day bucket, where some of it will live for
longer than it would have otherwise, but we avoid having to traverse the
entire trash hierarchy until it's time to clean it up. This way, the
migration only takes three filesystem write operations.

Refs: https://github.com/storj/storj/issues/6609
Change-Id: I7b6b8f84b3c44eca9cc18aef705e5872a81f351f